Add support for PGO instrumentation
This commit adds support for PGO instrumentation. This should be
enabled by adding the `--pgo-instrument` option to the Gradle
command line. When this is done, then the generated binary will
be compiled with PGO instrumentation enabled, and the binary
name will be suffixed with `-instrumented`.

It is possible to run the instrumented binary directly too,
in which case the profile files will be written in the same
directory as the binary.